Move EmbeddedPaginationHelperImpl to persistence module.
